+++ kata2.spec.ts
@@ -1,78 +1,67 @@
-import { ParsingError } from "./kata1";
-import { PrimitiveParser } from "./kata2";
+import { BaseParser, ParsingError } from "./api";
 
 
-describe("#2 Kata - Use Expressions to implement primitives", () => {
-  describe("Numbers", () => {
-    it("should parse a digit", () => {
-      // Given a parser with some input containing integers
-      const parser = new PrimitiveParser();
-      parser.setInput("1234");
-      // When the parser tries to consume a single digit
-      const output = parser.parseDigit();
-      // Then it should have consumed the first digit
-      expect(output).toBe("1");
-      expect(parser.getCurrent()).toBe("2");
-    });
-
-    it("should parse decimals", () => {
-      // Given a parser with some input containing integers
-      const parser = new PrimitiveParser();
-      parser.setInput("1234");
-      // When the parser tries to consume a decimal number with
-      // multiple digits
-      const output = parser.parseDecimal();
-      // Then it should have parsed the whole input
-      expect(output).toEqual(1234);
-    });
-
-    it("should parse hex-decimals", () => {
-      // Given a parser with some input containing a hex number
-      const parser = new PrimitiveParser();
-      parser.setInput("0xFF");
-      // When the parser tries to consume a decimal number with
-      // multiple digits
-      const output = parser.parseHexDecimal();
-      // Then it should have parsed the whole input
-      expect(output).toEqual(255);
-    });
-
-    it("should parse binary numbers", () => {
-      // Given a parser with some input containing a binary number
-      const parser = new PrimitiveParser();
-      parser.setInput("0b101010");
-      // When the parser tries to consume a decimal number with
-      // multiple digits
-      const output = parser.parseBinary();
-      // Then it should have parsed the whole input
-      expect(output).toEqual(42);
-    });
-
-    it("should parse floating point numbers", () => {
-      // Given a parser with some input containing a float number
-      const parser = new PrimitiveParser();
-      parser.setInput("0.55");
-      // When the parser tries to consume a decimal number with
-      // multiple digits
-      const output = parser.parseFloat();
-      // Then it should have parsed the whole input
-      expect(output).toEqual(0.55);
-    });
-
-    it("should parse all supported number types", () => {
-      const parser = new PrimitiveParser();
-      // Decimals
-      parser.setInput("55");
-      expect(parser.parseNumber()).toBe(55);
-      // Hex decimals
-      parser.setInput("0xf");
-      expect(parser.parseNumber()).toBe(15);
-      // Binary numbers
-      parser.setInput("0b101010");
-      expect(parser.parseNumber()).toBe(42);
-      // Floats
-      parser.setInput("0.55");
-      expect(parser.parseNumber()).toBe(0.55);
+describe("#2 Kata - Matching Expressions", () => {
+  describe("zeroOrMore", () => {
+    it("should capture as much of the input as possible", () => {
+      // Given a base parser with some input
+      const parser = new BaseParser();
+      parser.setInput("1234abcd");
+      // And given a function that
+      const integers = () => {
+        // Check if the current char is a digit and move
+        // the cursor forward if so and move on
+        const current = parser.getCurrent();
+        if (/\d/.test(current)) {
+          parser.next();
+          return current;
+        }
+        // Stop the loop at the end of the sequence of integers
+        parser.error('Not Int');
+      };
+      // When zero or more integers are consumed
+      const output = parser.zeroOrMore(integers);
+      // Then it should consume all the input since it all matches
+      expect(output).toEqual(['1', '2', '3', '4']);
     });
   });
+
+  describe("or", () => {
+    it("should choose the first expression that succeeds", () => {
+      // Given a new base parser with some input
+      const parser = new BaseParser();
+      parser.setInput("abacate");
+      // When faced with a list of options, where just the last one
+      // matches successfuly
+      const output = parser.or([
+        () => parser.expectString("abelha"),
+        () => parser.expectString("abada"),
+        () => parser.expectString("abacate"),
+      ]);
+      // Then it should match the successful one
+      expect(output).toBe("abacate");
+    });
+
+    it("should fail if no option succeeds", () => {
+      // Given a new base parser with some input
+      const parser = new BaseParser();
+      parser.setInput("abacate");
+      // When faced with a list of options
+      expect(() => parser.or([() => parser.error("a"), () => parser.error("b")]))
+        .toThrow(new ParsingError('No option found'));
+    });
+  });
+
+  describe("oneOrMore", () => {
+  });
+
+  describe("optional", () => {
+  });
+
+  describe("not", () => {
+  });
+
+  describe("and", () => {
+  });
+
 });
